AT12 - only allow selection of mp3 and mp4

+++ src/main/java/transcription/services/ServiceDownload.java
@@ -26,6 +26,9 @@ public class ServiceDownload {
             if (json.length()>34 && json.substring(1,34).equals("java.nio.file.NoSuchFileException")){
                 return "File is not transcribed yet.";
             }
+            if (json.equals("\"File does not exist\"")){
+                return "Job name does not exist";
+            }
             Files.write( Paths.get(filePath+"/"+job+".txt"), providerDownload.executeDownloadHttp(job + ".json").getBytes());
             LogManager.getLogManager().getLogger(Logger.GLOBAL_LOGGER_NAME).setLevel(Level.INFO);
             LOGGER.info("complete");
